imporved text report

+++ DuplicateCheck/src/com.dupcheck.strategy/com/dupcheck/strategy/DeleteFileStrategy.java
@@ -8,6 +8,7 @@ import java.nio.file.Path;
 public class DeleteFileStrategy implements FeatureStrategy{
 	
 		@Override public void execute(List<File> list){
+			if(list != null){
 				list.forEach(e -> {
 					try{
 						Files.deleteIfExists(e.toPath());
@@ -16,6 +17,9 @@ public class DeleteFileStrategy implements FeatureStrategy{
 						System.out.println("Error Occured when deleting files.");
 					}
 				});
+			}else{
+				System.out.println("Null argument is passed.");
+			}
 		}
 	
 }
